ifctester: split shared test file per fixture
test_fixtures.py is shared with our other open ids-fixture PRs
(#371-branch, #9203), which each independently create the same file
and collide with each other and with this branch. Splitting into
one file per fixture avoids the clash; fixture folders under
test/fixtures/ never collide since each PR uses its own subfolder.
