+++ src/ifctester/test/test_fixture_attribute_optional_unset.py
@@ -45,17 +45,3 @@ class TestAttributeOptionalUnsetFixture:
         spec = specs.specifications[0]
         assert spec.status is True
         assert spec.requirements[0].status is True
-
-
-class TestAttributeLogicalUnknownFixture:
-    def test_prohibited_attribute_fails_when_set_to_logical_unknown(self):
-        # IfcPresentationLayerWithStyle.LayerOn is an IFCLOGICAL set to the
-        # tri-state value UNKNOWN, which is a genuinely present value, not
-        # an absent one. A cardinality="prohibited" requirement on it must
-        # fail, since the attribute is provided.
-        specs = ids.open(os.path.join(FIXTURES, "attribute_logical_unknown", "attribute_logical_unknown.ids"))
-        ifc = ifcopenshell.open(os.path.join(FIXTURES, "attribute_logical_unknown", "attribute_logical_unknown.ifc"))
-        specs.validate(ifc)
-        spec = specs.specifications[0]
-        assert spec.status is False
-        assert spec.requirements[0].status is False
